+++ tests/fakeparser.py
@@ -64,6 +64,7 @@ class Parser:
         raise NotImplementedError
 
     parallel = reader._parser.Parser.parallel
+    persistent_session = staticmethod(nullcontext)
 
     def retrieve(self, url, http_etag, http_last_modified, is_parallel):
         return nullcontext(RetrieveResult(BytesIO(b'opaque')))